separate `get` mdoule for getting accounts
