+++ library/observables-connect/src/accounts/util.ts
@@ -3,68 +3,14 @@ SPDX-License-Identifier: GPL-3.0-only */
 
 import type {
   ExtensionAccount,
-  ExtensionEnableResults,
   HandleImportExtension,
   ImportedAccount,
 } from '@w3ux/types'
 import { formatAccountSs58, isValidAddress } from '@w3ux/utils'
-import { DefaultHandleImportExtension, DefaultSS58 } from '../consts'
+import { DefaultHandleImportExtension } from '../consts'
 import { getActiveAccountLocal, getLocalExternalAccounts } from './local'
 
-// Connects to the provided extensions and fetches their accounts
-export const getAccountsFromExtensions = async (
-  extensions: ExtensionEnableResults
-): Promise<ExtensionAccount[]> => {
-  try {
-    const results = await Promise.allSettled(
-      Array.from(extensions.values()).map(({ extension }) =>
-        extension.accounts.get()
-      )
-    )
-    const extensionEntries = Array.from(extensions.entries())
-    const accounts: ExtensionAccount[] = []
-
-    for (let i = 0; i < results.length; i++) {
-      const result = results[i]
-      const source = extensionEntries[i][0]
-      const signer = extensionEntries[i][1].extension.signer
-
-      if (result.status === 'fulfilled') {
-        const filtered = result.value
-          // Reformat addresses with default ss58 prefix
-          .map((account) => {
-            const address = formatAccountSs58(account.address, DefaultSS58)
-            if (!address) {
-              return null
-            }
-            return {
-              ...account,
-              address,
-            }
-          })
-          // Remove null entries resulting from invalid formatted addresses
-          .filter((a) => a !== null)
-          // Remove accounts that have already been imported
-          .filter(({ address }) => !accounts.find((a) => address === a.address))
-          // Reformat entries to include extension source
-          .map(({ address, name }) => ({
-            address,
-            name,
-            source,
-            signer,
-          }))
-
-        accounts.push(...filtered)
-      }
-    }
-    return accounts
-  } catch (e) {
-    console.error('Error during enable and format call: ', e)
-    return []
-  }
-}
-
-// Gets accounts to be imported and commits them to state.
+// Gets accounts to be imported and commits them to state
 export const handleExtensionAccountsUpdate = (
   id: string,
   currentAccounts: ExtensionAccount[],
